#134 refactor process of indexing-progress-monitoring

+++ core/src/main/java/org/hibernate/search/jsr352/internal/steps/lucene/StepContextData.java
@@ -37,7 +37,7 @@ public class StepContextData implements Serializable {
 		partitionProgress = new PartitionProgress( partitionID, entityName );
 	}
 
-	public void documentAdded(long increment) {
+	public void documentAdded(int increment) {
 		partitionProgress.documentsAdded( increment );
 	}
